requirements.txt
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
cryptography==3.0
decorator==4.4.2
google-api-core==1.22.0
google-auth==1.20.1
google-cloud-core==1.4.1
google-cloud-translate==3.0.1
google-cloud-vision==1.0.0
googleapis-common-protos==1.52.0
grpcio==1.31.0
idna==2.10
libcst==0.3.9
mypy-extensions==0.4.3
proto-plus==1.7.0
protobuf==3.12.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
python-telegram-bot==12.8
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
rsa==4.6
six==1.15.0
tornado==6.0.4
typing-extensions==3.7.4.2
typing-inspect==0.6.0
urllib3==1.25.10
